@@ -4,13 +4,16 @@
 	"description": "Create a lazily evaluated value",
 	"license": "MIT",
 	"repository": "sindresorhus/lazy-value",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -24,21 +27,19 @@
 		"value",
 		"lazily",
 		"laziness",
-		"val",
 		"evaluation",
 		"eval",
 		"execute",
 		"getter",
 		"function",
-		"fn",
 		"memoize",
 		"cache",
 		"defer",
 		"deferred"
 	],
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"tsd": "^0.14.0",
+		"xo": "^0.38.2"
 	}
 }
